Options for the projection solves: - PRECON   (for both projections; without it nothing works)
Best option for solving the second projection at this stage:
- PRECON_DIAGONAL_2ND_PROJ   (decisively more efficient than the current version of  PRECON_VERTICAL_COLUMN_2ND_PROJ; a combination of both may be the solution)
- CONTROL_PRECONDITIONED_RESIDUAL_PROJ2    (in the second projection this is definitely beneficial)

Best option for the first projection
- PRECON_DIAGONAL_2ND_PROJ

Note that   CONTROL_PRECONDITIONED_RESIDUAL_PROJ1   is NOT used. With it, the code does not work at all.
Since this option is, however, definitely beneficial in the second project, there should be a bug somewhere in the
implementation for the first projection. This I need to hunt down next. Maybe the 1st projection is the culprit after all.