Since October 24 I had worked on various variants of handling the nodal pressure evolution so as to move it as close as possible to a the behaviour of a symplectic integrator. This was only partially successful so far, and in the meantime I lost track of all the changes I had made.   This commit is a re-implementation of the working version of October 24, which successfully handled the Straka and internal wave tests in psinc and comp modes. I have labelled all code pieces where re-coding had to take place by
            /* TODO: controlled redo of changes from 2018.10.24 to 2018.11.11 */

so we can easily find these points and analyse the influence of each of these
changes.

The current version should be formally second order, but it does damp acoustic
modes. Therefore, the 1D acoustic test will not come out all too successfully.
As a starting point for the vertical slice developments in A02 it should be valuable,
however.