fixed bug in divergence nodes Pv cell to node averaging
on the top and bottom most rows the cell to node averaging averages to the flux across the cell interface, not the across the node. this is a very dirty fix to that problem that does the cell to node averaging twice. I will have to clean this up eventually.