second_projection.py: Fixed sign errors in euler forward and a few minor tweaks, see details
While I fixed the signs of the coefficients in euler backwards and correction nodes, I forgot to fix it in the euler forward. Now that has been fixed. I also moved the coefficients to be more readable in the u,v,w order. Finally, I really should decide on which approach to implement the divergence nodes.