fixed typo in the unstable lamb test
so the initial 30 time steps or so are very close to the 0.4.0 release, but the solutions diverge after about 50 time steps due to build up in this difference and the sensitivity of the instabilities. However, the amplitudes are slightly larger in our current run and 0.4.0 slightly underestimated the instability growth rate. So we might actually be on the better side of things here. Nevertheless, may be worth looking into this in the future.